TCS- Backend/requirements.txt
pymongo[srv]==3.12.0
pydantic==1.8.2
fastapi==0.70.1
PyJWT==2.3.0
bcrypt==3.2.0
requests==2.26.0
uvicorn==0.15.0
python-multipart==0.0.5